Generalize geometry collapsing algo and move functions
